requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
av==11.0.0
babel==2.16.0
bibtexparser==2.0.0b7
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
clients==1.5
clldutils==3.22.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.3.0
cryptography==43.0.3
csvw==3.3.1
ctranslate2==4.4.0
cycler==0.12.1
distro==1.9.0
dlinfo==1.2.1
docopt==0.6.2
einops==0.8.0
einops-exts==0.0.4
filelock==3.13.1
firebase-admin==6.5.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.2.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.7
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munch==4.0.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.3
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.52.1
optuna==4.0.0
pandas==2.2.3
phonemizer==3.3.0
pillow==10.2.0
pooch==1.8.2
primePy==1.3
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pylatexenc==2.10
pyparsing==3.1.4
pysrt==1.1.2
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.2
PyYAML==6.0.2
rdflib==7.0.0
referencing==0.35.1
requests==2.32.3
rfc3986==1.5.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
segments==2.2.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.1
SQLAlchemy==2.0.35
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.15.2
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchmetrics==1.4.3
tqdm==4.66.5
transformers==4.39.3
triton==3.1.0
typer==0.12.5
typing==3.7.4.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.14.0
pyloudnorm
replicate